+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2021-12-30
+### Added
+- Hunt! feature that let you search search all potential indicators present in the web page visible in the current active tab of the browser.
+- Trash bin button to clear the input text field.
+- Open in new/current tab button
+
+### Changed
+- Behaviour of indicators text selection. Now only submitted indicator will be saved inside the input text field of the popup.
+- Buttons icons inside the input text field.
+- Settings popup appearance
+
+### Fixed
+- Behaviour of the domain extraction feature icon in the textfield
+
 ## [0.5.2] - 2021-11-16
 ### Added
 - Auto-submit feature for CentralOps resource (domain whois, network whois, dns)